% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_functions.R
\name{run_gsea}
\alias{run_gsea}
\title{Performs gene set enrichment analysis with fgsea on a table of gene stats}
\usage{
run_gsea(gene_stats, term2gene, gene_var = "Gene", rank_var = "logFC",
  dir = "both", method = "fgsea", min_size = 1, max_size = 500)
}
\arguments{
\item{gene_stats}{a table of gene level stats.}

\item{term2gene}{a table of gene_sets in term2gene format.}

\item{gene_var}{name of the column containing gene names.}

\item{rank_var}{name of the column to rank genes by.}

\item{dir}{direction(s) to consider. Options are ("both","pos","neg")}

\item{method}{GSEA methods. Options are ("fgsea").}

\item{min_size}{minimal size of a gene set to test.}

\item{max_size}{maximal size of a gene set to test. Defaults to 500.}
}
\value{
A table with the following columns:
\describe{
term: gene set name. \cr
p_value: enrichment p-value. \cr
p_adjust: BH-adjusted p-value. \cr
ES: enrichment score. \cr
NES: enrichment score normalized to mean enrichment of random samples of the same size. \cr
direction: enrichment direction. \cr
size: size of the gene set. \cr
leading_edge: vector with indexes of leading edge genes that drive the enrichment. \cr
}
}
\description{
Performs gene set enrichment analysis with fgsea on a table of gene stats
}
