% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_functions.R
\name{run_hyper}
\alias{run_hyper}
\title{Performs over representation analysis with fisher's exact test on a gene list or table of gene stats}
\usage{
run_hyper(
  gene_stats,
  term2gene,
  gene_var = "Gene",
  rank_var = "logFC",
  universe = NULL,
  n_genes = 100,
  dir = "both",
  p_adjust_method = "BH",
  min_size = 1,
  max_size = Inf
)
}
\arguments{
\item{gene_stats}{a small list of significant genes or table of gene level stats.}

\item{term2gene}{a table of gene_sets in term2gene format.}

\item{gene_var}{if gene_stats is table, name of the column containing gene names.}

\item{rank_var}{if gene_stats is table, name of the column containing gene stats to rank by.}

\item{universe}{an optional list of genes that represent the universe. Defaults to genes if gene_stats is a table.
Defaults to all genes if gene_stats is a list.}

\item{n_genes}{if gene_stats is table, number of genes to select after ranking by rank_var.}

\item{dir}{if gene_stats, direction(s) to consider. Options are ("both","pos","neg").}

\item{p_adjust_method}{Method for p-value adjustment. Defaults to BH}

\item{min_size}{minimal size of a gene set to test.}

\item{max_size}{maximal size of a gene set to test.}
}
\value{
A table with the following columns:
\describe{
term: gene set name. \cr
p_value: enrichment p-value. \cr
p_adjust: adjusted p-value. \cr
odds_ratio: the odds ratio the genes being in the set vs not in the set. \cr
direction: enrichment direction. \cr
size: size of the gene set. \cr
overlap_size: size of the overlap between the gene set and significant genes list. \cr
overlap: genes in both the gene set and significant genes list. \cr
}
}
\description{
Performs over representation analysis with fisher's exact test on a gene list or table of gene stats
}
